+++ blockfrost-api/CHANGELOG.md
@@ -5,6 +5,7 @@
   * `redeemerCount` field to `Transaction`
   * `dataHash` field to `UtxoInput` and `UtxoOutput`
   * `outputIndex` field to `UtxoOutput`
+  * `script` field to `AddressInfo`
 
 # Version [0.1.0.0](https://github.com/blockfrost/blockfrost-haskell/compare/initial...0.1.0.0) (2021-09-14)